arm64: Simplify efi_mem_init
Reduce the EFI mem_map loop to only setting flags and finding the
largest free memory region. Then, apply memregions_split() for
the code/data region split (which requires ensuring _etext is
page aligned). Finally, do the rest of the things that used to be
done in the EFI mem_map loop in a separate mem_region loop.

Reviewed-by: Nikos Nikoleris <nikos.nikoleris@arm.com>
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>