s390x: Use local accel variable in arch_cmd_s390x
By the time we end up in arch_cmd_s390x() the global ACCEL variable
has already been processed and is passed to arch_cmd_s390x() as an
argument. We should use the local argument to achieve the correct
behavior and not rely on global variables.

Fix this by changing ACCEL to lower case.

Reported-by: shellcheck SC2153
Reviewed-by: Andrew Jones <andrew.jones@linux.dev>
Acked-by: Janosch Frank <frankja@linux.ibm.com>
Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Link: https://lore.kernel.org/r/20240406122456.405139-2-npiggin@gmail.com
Message-ID: <20240406122456.405139-2-npiggin@gmail.com>
[frankja@linux.ibm.com: Improved commit message]
Signed-off-by: Janosch Frank <frankja@linux.ibm.com>