arm64: Default to 4K translation granule
Some arm64 implementations in the wild, like the Apple parts, do not
support the 64K translation granule. This can be a bit annoying when
running with the defaults on such hardware, as every test fails
before getting the MMU turned on.

Switch the default page size to 4K with the intention of having the
default setting be the most widely applicable one.

Signed-off-by: Oliver Upton <oliver.upton@linux.dev>
Acked-by: Zenghui Yu <zenghui.yu@linux.dev>
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>