shellcheck: Suppress SC2209 quoting warning in config.mak
It's not necessary to quote strings in simple command variables like
this where the pattern makes the intention quite clear.

config.mak is also included as Makefile, and in that case the quotes
do slightly change behaviour (the quotes are used when invoking the
command), and is not the typical Makefile style.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Message-ID: <20240503052510.968229-3-npiggin@gmail.com>
Tested-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>