travis.yml: Update the aarch64 and ppc64le jobs to Jammy
Ubuntu Focal is now four years old, so we might miss some new compiler
warnings that have been introduced in later versions. Ubuntu Jammy is
available in Travis since a while already, so let's update to that
version now.

Unfortunately, there seems to be a linking problem with Jammy on s390x,
so we have to keep the s390x on Focal for now.

Message-ID: <20240507133426.211454-1-thuth@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>