x86: vmexit: Allow IPI test to be accelerated by SVM AVIC
The vmexit_ipi test can be used as a rough benchmark for IPI performance
since commit 8a8c1fc3b1f8 ("vmexit: measure IPI and EOI cost") added
reporting of the average number of cycles taken for IPI delivery. Avoid
exposing a PIT to the guest so that SVM AVIC is not inhibited and IPI
acceleration can be tested when available and enabled by the host.

Signed-off-by: Alejandro Jimenez <alejandro.j.jimenez@oracle.com>
Link: https://lore.kernel.org/r/20240522215755.197363-1-alejandro.j.jimenez@oracle.com
Signed-off-by: Sean Christopherson <seanjc@google.com>