Bump Fedora version to 40
Version 37 is out of date, let's use a supported one instead.

The "xsave" test now tries to execute more tests with TCG, but one
of them is failing with the QEMU from Fedora 40, so we have to
disable the xsave test for now.
The problem will be fixed in QEMU 9.1 with this patch here:
https://lore.kernel.org/qemu-devel/20240603100405.619418-1-pbonzini@redhat.com/

And there is also an additional problem with the "realmode" test. As
diagnosed by Paolo: "It turns out that with some versions of clang,
realmode.flat has become big enough that it overlaps the stack used
by the multiboot option ROM loader. The result is that a couple
instructions are overwritten." Thus disable the realmode test in the
Clang build until the problem is fixed.

Acked-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>
Message-ID: <20240607065147.63226-1-thuth@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>