riscv: Track memory above 3G
We chop the main memory region (which gets identity mapped) off at
the 3G boundary because the virtual memory allocator claims the 3G-4G
address range. Unit tests may want to be able to access the higher
memory, though, so keep track of it by putting it in its own memory
region. Since the test framework isn't using that memory region, flag
it as unused so unit tests will be confident that they can use the
memory for whatever they like.

Signed-off-by: Andrew Jones <andrew.jones@linux.dev>