x86: Force host-phys-bits for normal maxphyaddr access tests
Explicitly force host-phys-bits for the access tests that intends to run
with the host's MAXPHYADDR, as QEMU only forces host-phys-bits for "max"
CPUs as of version 6.0 (see commit 5a140b255d ("x86/cpu: Use max host
physical address if -cpu max option is applied")).

Running the access test with an older QEMU, e.g. 5.3, on a CPU with
MAXPHYADDR=52 and allow_smaller_maxphyaddr=N (i.e. with TDP enabled)
fails miserably as the test isn't aware that bit 51 is a legal physical
address bit.

Amusingly, the KVM-Unit-Test appears to pre-date the QEMU change by a few
months, which suggests that the KVM-Unit-Test change was made without
proper testing, and the QEMU change was made to fix the problem.

Fixes: 11cec501 ("x86: cover emulation of reduced MAXPHYADDR")
Link: https://lore.kernel.org/r/20240823221833.2868-1-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>