riscv: QEMU Sstc timer stop workaround
QEMU doesn't stop a pending timer when UINT64_MAX is written to
stimecmp, see QEMU commit ae0edf2188b3 ("target/riscv: No need to
re-start QEMU timer when timecmp == UINT64_MAX"). A fix has been
posted[1], but we need a solution in kvm-unit-tests anyway in order
to support older QEMU versions. A bit of an ugly workaround is to
simply subtract one from UINT64_MAX, which is still a really big
number, but not the exact number QEMU is using to decide it should
skip the timer update.

Link: https://lore.kernel.org/all/20240829084002.1805006-2-ajones@ventanamicro.com/ [1]
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>