riscv: sbi: Rename __*_sbi_ecall to sbi_*
Let's name all SBI ecall wrappers with the same pattern of sbi_<ext>
and sbi_<ext>_<func>, even those that are local to the SBI tests.
This is good for consistency and if those functions are ever promoted
to the library then we won't have to change all their invocations.

Signed-off-by: Andrew Jones <andrew.jones@linux.dev>