pmu_lbr: drop check for MSR_LBR_TOS != 0
While this is not likely, it is valid for the MSR_LBR_TOS
to contain 0 value, after a test which issues a series of branches, if the
number of branches recorded was divisible by the number of LBR msrs.

This unfortunately depends on the compiler, the number of LBR registers,
and it is not even deterministic between different runs of the test,
because interrupts, rescheduling, and various other events can affect total
number of branches done.

Therefore drop the check, instead of trying to fix it.

Signed-off-by: Maxim Levitsky <mlevitsk@redhat.com>
Link: https://lore.kernel.org/r/20241002235658.215903-1-mlevitsk@redhat.com
Signed-off-by: Sean Christopherson <seanjc@google.com>