s390x: Add library functions for exiting from snippet
It is useful to be able to force an exit to the host from the snippet,
as well as do so while returning a value.
Add this functionality, also add helper functions for the host to check
for an exit and get or check the value.
Use diag 0x44 and 0x9c for this.
Add a guest specific snippet header file and rename snippet.h to reflect
that it is host specific.

Signed-off-by: Nina Schoetterl-Glausch <nsg@linux.ibm.com>
Link: https://lore.kernel.org/r/20241016180320.686132-5-nsg@linux.ibm.com
[ nrb: fix out-of-tree builds ]
Signed-off-by: Nico Boehr <nrb@linux.ibm.com>