nVMX: Account for gaps in fixed performance counters
Update the nVMX PERF_GLOBAL_CTRL test to play nice with PMUs that have
a discontiguous set of PMCs.  On CPUs with PMU Version 5 or later, the
set of fixed PMCs may not be contiguous.  Use the logic recommended by
the Intel SDM to determine if a Fixed Counter is supported:

	FxCtr[i]_is_supported := ECX[i] || (EDX[4:0] > i);

For example, it's perfectly valid to have CPUID.0AH.EDX[4:0] == 3 and
CPUID.0AH.ECX == 0x77, but checking the fixed counter index against
CPUID.0AH.EDX[4:0] only, will deem that FxCtr[6:4] are not supported.

Opportunistically add anythread_deprecated to cpuidA_edx.

Signed-off-by: Zide Chen <zide.chen@intel.com>
Reviewed-by: Zhenyu Wang <zhenyuw@linux.intel.com>
Link: https://lore.kernel.org/r/20241118225207.16596-1-zide.chen@intel.com
[sean: massage changelog, keep "counters" in the names]
Signed-off-by: Sean Christopherson <seanjc@google.com>