riscv: sbi: Improve gen_report
Make several improvements to gen_report(), starting by renaming it
to something less generic (sbiret_report) and then, instead of
relying on report prefix to link the report_info with the unexpected
return values to the failed test, use the test string itself, by
taking it as a parameter. And, reporting sbiret.error and
sbiret.value separately was more verbose than necessary since the
report_info can be used to see which one failed, so combine them.
Finally, return the pass/fail result of the test in case a caller
wants to use it.

Tested-by: Clément Léger <cleger@rivosinc.com>
Reviewed-by: Clément Léger <cleger@rivosinc.com>
Signed-off-by: Andrew Jones <andrew.jones@linux.dev>